backend/reports/report ex/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.0,<5.0

# ── Async task queue ───────────────────────────────────────────────────────
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0

# ── Payment gateways ───────────────────────────────────────────────────────
stripe>=9.0,<10.0
requests>=2.31,<3.0

# ── Report generation ──────────────────────────────────────────────────────
reportlab>=4.0,<5.0
openpyxl>=3.1,<4.0
